Some refactoring to allow code reuse for the next parts of this example
1. Removed `employee.c` and moved its content plus `set_read_only()` to
`employee.h;
2. updated `Makefile` to reflect the changes.
